# swagger.yaml
swagger: "2.0" # https://stackoverflow.com/a/59749691/14797322
info:
  title: "magiclinks.dev"
  description: "The v2 API specification for the magiclinksdev project.
  \n\n
  The default JWK Set relative URL path is `/api/v2/jwks.json`.
  \n\n
  The documentation site is at https://docs.magiclinks.dev
  \n
  This is an Apache License 2.0 project: https://github.com/MicahParks/magiclinksdev
  \n
  The optional SaaS platform's landing page is: https://magiclinks.dev
  "
  version: "2.0.0"
host: "magiclinks.dev"
basePath: "/api/v2"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /magic-link/create:
    post:
      summary: "Create a magic link."
      operationId: "magicLinkCreate"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/MagicLinkCreateRequest"
      responses:
        201:
          description: "The magic link was created."
          schema:
            $ref: "#/definitions/MagicLinkCreateResponse"
        default:
          description: "An unexpected error occurred."
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "A copy of the HTTP response status code."
        format: "int32"
      message:
        type: "string"
        description: "A human-readable error message."
      requestMetadata:
        $ref: "#/definitions/RequestMetadata"
  RequestMetadata:
    type: "object"
    properties:
      uuid:
        type: "string"
        description: "A UUID that uniquely identifies the request."
        format: "uuid"
...
  JWTCreateParams:
    description: "Parameters used to create a JWT."
    type: "object"
    properties:
      alg:
        description: "The algorithm to use when signing the JWT. It defaults depends on the server's configuration. The
        default server configuration is \"EdDSA\". The default server options are \"EdDSA\" and \"RS256\"."
        type: "string"
      claims:
        description: 'Any valid JSON object. Do not provide any JSON attributes mentioned in RFC 7519 section 4.1 as
        this will cause an error. These are JWT "Registered Claim Names".'
        type: "object"
        externalDocs:
          description: "RFC 7519 section 4.1"
          url: "https://tools.ietf.org/html/rfc7519#section-4.1"
      lifespanSeconds:
        description: "The lifespan of the JWT in seconds. The JWT's lifespan starts after creation. For magic links,
        after the magic link has been visited. It defaults to 5 minutes. The minimum value is 5 seconds and the maximum
        value is 7905600000 seconds, which is a bit over 250 years."
        default: 300
        type: "integer"
...
  MagicLinkCreateParams:
    description: "Parameters to create a magic link."
    type: "object"
    properties:
      jwtCreateParams:
        $ref: "#/definitions/JWTCreateParams"
      lifespanSeconds:
        description: "The lifespan of the magic link in seconds. The magic link's lifespan starts after it has been
        created. It defaults to 1 hour. The minimum value is 5 seconds and the maximum value is 7905600000 seconds,
        which is a bit over 250 years."
        type: "integer"
        default: 3600
      redirectQueryKey:
        description: 'The URL query key in the redirectURL to contain the signed JWT when the magic link is used. By
        default, "jwt" is used.'
        type: "string"
      redirectURL:
        description: "The URL to redirect to with the signed JWT when the magic link is used."
        type: "string"
    required:
      - "redirectURL"
  MagicLinkCreateRequest:
    type: "object"
    properties:
      magicLinkCreateParams:
        $ref: "#/definitions/MagicLinkCreateParams"
    required:
      - "magicLinkCreateParams"
  MagicLinkCreateResults:
    type: "object"
    properties:
      magicLink:
        description: "The URL that will act as a magic link. When this URL is visited, a new JWT will be created. A
        redirect wil be performed with this new JWT in the redirect URL's query parameter."
        type: "string"
      secret:
        description: "The secret embedded in the magic link."
        type: "string"
  MagicLinkCreateResponse:
    type: "object"
    properties:
      magicLinkCreateResults:
        $ref: "#/definitions/MagicLinkCreateResults"
      requestMetadata:
        $ref: "#/definitions/RequestMetadata"
...
securityDefinitions:
  apiKey:
    type: "apiKey"
    in: "header"
    name: "X-API-KEY"
security:
  - apiKey: [ ]
...